requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
